LayoutTests/fast/css-grid-layout/grid-repeat-calc-expected.txt

Tests using calculated value in function() for 'grid-template-rows' CSS property works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['grid-template-rows'] is ""
testDiv.style['grid-template-rows'] = 'repeat(calc(1 + 1), 18px)'
PASS testDiv.style['grid-template-rows'] is "18px 18px"
PASS window.getComputedStyle(testDiv).getPropertyValue('grid-template-rows') is "18px 18px"
PASS successfullyParsed is true

TEST COMPLETE
